Split etl test file
